add fix

+++ src/server/server_protocol.cpp
@@ -178,7 +178,7 @@ CommandDTO ServerProtocol::receive_buy_weapon_request() {
 
 CommandDTO ServerProtocol::receive_buy_ammo_request() {
     uint8_t primary_code = this->receive_byte();
-    uint8_t ammo = this->receive_big_endian_number();
+    uint16_t ammo = this->receive_big_endian_number();
     bool primary = false;
     if (primary_code == CODE_CHOOSE_PRIMARY) {
         primary = true;
